shifts any2vec callback under gensim/models
